+++ example/src/comps/counter/counter.tsx
@@ -1,5 +1,6 @@
 import { $computed } from 'frame3'
 import { $size, setSize } from '../../store/size'
+import { Flex } from '../flex/flex'
 import './counter.css'
 
 export const Counter = () => {
@@ -8,15 +9,15 @@ export const Counter = () => {
   return (
     <>
       {$outOfBound && <span className="outOfBound">out of bound</span>}
-      <div className={`center container ${$outOfBound && 'outOfBound'}`}>
-        <div className="center operator" onClick={() => !$outOfBound && setSize($size - 1)}>
+      <Flex className={`container ${$outOfBound && 'outOfBound'}`}>
+        <Flex className="operator" onClick={() => !$outOfBound && setSize($size - 1)}>
           -
-        </div>
-        <div className="center shower">{$size}</div>
-        <div className="center operator" onClick={() => setSize($size + 1)}>
+        </Flex>
+        <Flex className="shower">{$size}</Flex>
+        <Flex className="operator" onClick={() => setSize($size + 1)}>
           +
-        </div>
-      </div>
+        </Flex>
+      </Flex>
     </>
   )
 }