+++ example/src/index.ts
@@ -1,4 +1,8 @@
-import { mount } from 'frame3'
+import { defineTheme, mount } from 'frame3'
 import { App } from './app'
+
+const theme = defineTheme({
+  color: 'blue',
+})
 
 mount(App, document.body)