+++ packages/vite-plugin/CHANGELOG.md
@@ -1,5 +1,13 @@
 # vite-plugin-frame3
 
+## 1.1.6
+
+### Patch Changes
+
+- 添加配置 css 主题 api 以及修复嵌套组件渲染异常
+- Updated dependencies
+  - frame3-compiler-sfc@1.1.5
+
 ## 1.1.5
 
 ### Patch Changes